Got tests green, fixed setting transforms

+++ src/window_size.rs
@@ -6,8 +6,8 @@ use bevy::{prelude::*, render::camera::Camera};
 /// Hopefully this may be integrated into bevy in future.
 #[derive(Default)]
 pub struct WindowSize {
-    pub height: u32,
-    pub width: u32,
+    pub height: f32,
+    pub width: f32,
 }
 
 /// Syncs the window width to the camera
@@ -18,8 +18,8 @@ pub fn window_size(windows: Res<Windows>, mut camera_query: Query<(&Camera, &mut
             // Note: This conversion will be lossy, but u32 is used internally 
             // for texture calcs.
             let windowsize = WindowSize {
-                height: window.height() as u32, 
-                width: window.width() as u32,
+                height: window.height(), 
+                width: window.width(),
             };
             *size = windowsize;
         }
